do not verify pgp key for tor browser
This is lame, but on a fresh install the user will not have the key in
their keyring. We could add it, but doing that blindly seems even worse
than just skipping it. This way the user will at least get an error
whenever they try to upgrade the package, which will prompt them to
acquire and verify the key manually.