disable msmtpq connectivity test
Netcat (and ping) are not available in the mutt jail, so calling msmtpq
(from mailsync.sh) from within that jail throws an error. But
mailsync.sh does its own connectivity testing, so we don't actually need
this!

When offline, mail will still be queued. When online, mail will still be
sent.