install pulseaudio
Existing users will want to remove the old udev rule for device
switching, and the /etc/asound.conf file it generated.

    # rm /etc/udev/rules.d/00-local.rules
    # rm /etc/asound.conf

A new /etc/asound.conf is provided by the pulseaudio-alsa package. If
our old file is not removed prior to installation, the new file will be
installed at /etc/asound.conf.pacnew. This will need to be moved to the
proper location.