install hostsctl from aur
Prior to running this, existing users should put any custom host
definitions into /etc/hostsctl.d/10-hosts. This is due to a significant
refactor of hostsctl that was made after it was originally added to
spark. If you had not previously installed hostsctl, this is not
necessary.

    # mkdir /etc/hostsctl.d
    # echo "127.0.0.1" > /etc/hostsctl.d/10-hosts

The old script should also be deleted.

    # rm /usr/local/bin/hostsctl