run git-annex as a user service
When this role was created, Ansible did not support sytemd user units,
so we setup git-annex as a system service running as the user.  The only
practical difference in moving to a user service is that the user can
control the service without sudo, but we might as well.

Existing users should first stop, disable, and remove the old system
service:

    $ sudo systemctl stop git-annex@$USER.service
    $ sudo systemctl disable git-annex@$USER.service
    $ sudo rm /etc/systemd/system/git-annex@.*

If the system service is present in the trusted unit file, remove it.