run mailsync as a user service
When this role was created, Ansible did not support systemd user units,
so we setup mailsync as a system service and timer running as the user.
The only practical difference in moving to a user service is that the
user can control the service without sudo, but we might as well.

Existing users should first stop, disable, and remove the old system timer and
service:

    $ sudo systemctl stop mailsync@$USER.timer
    $ sudo systemctl disable mailsync@$USER.timer
    $ sudo rm /etc/systemd/system/mailsync@.*

If the system timer is present in the trusted unit file, remove it.