support kernel parameters
Note that this is only supported for grub. That is the only bootloader I
use. Patches for other bootloaders will be accepted.

The default params:

    * quiet the boot process
    * cause all virtual terminals to blank (ie, screen power off) after 60 seconds
    * enable panel self refresh v2

Blanking the screen is useful with physlock. Physlock does not run in X,
so DPMS settings have no effect on it.

PSR can potentially improve battery life on some machines, but cause
screen flicker on others. I have happily been using v2 on a ThinkPad
X260 for 6 months, but users should experiment between 1/2/3/4 to find
the best setting for their machine, or disable it entirely with 0. See:

* https://hansdegoede.livejournal.com/18653.html
* https://patchwork.kernel.org/patch/8182841/