support forwarding LAN zones to local LAN DNS
With this option enabled, if you connect to a local network which claims
to serve DNS for `mycorp.com`, unbound will forward DNS requests for
that domain to the local network DNS resolver.

If you connect to a local network which claims to serve DNS for a domain
like `mynetwork.lan` or `local.mesh`, and you have those domains
whitelisted in the `unbound.private_domains` var, you will be able to
successfully resolve those hostnames.

The `private_interfaces` option prevents forwarding the root zone to the
local network resolver. If the network claims to provide DNS for
`mycorp.com` and everything else, requests for `mycorp.com` will go to
the network's resolver, but requests for every other domain will still go
through unbound to the nameservers specified in the
`unbound.nameservers` var.