drop geteltorito
The AUR package currently has broken checksums. I can't remember the
last time I used this. And it is just a perl script, so there's no value
in having it as a package. Just curl the script if you need it, and add
it to $PATH if you're feeling ambitious.