set battery idle on framework machines
Apparently changes made via ectool only survive power off if the laptop
remains plugged in, so in practice we have to set this on every boot.

https://wiki.archlinux.org/title/Framework_Laptop_13#Battery_control

Setting the lower and upper charge percentage to the same value causes
the controller to idle the battery at that percentage. This seems like
it would be better for the long term health of the battery than giving
the battery a range to constantly dis/charge between. It also feels more
like what I'm used to from all my years of Thinkpads and TLP.

I have also set the max charge on my battery to 90% in the BIOS, as per:
https://guides.frame.work/Guide/Enabling+Battery+Saver+Functionality+in+BIOS/392